Cleaned up, commented, and standardized all of the files in the PHP folder

+++ php/coLoad.php
@@ -24,6 +24,7 @@ if (mysqli_connect_errno()) {
 	exit('Failed to connect to MySQL: ' . mysqli_connect_error());
 }
 
+// Don't really want to try and grab a variable that doesn't exist
 if (!isset($_SESSION['name'])) {
 	header('Location: ../index.html');
 	exit();
@@ -31,14 +32,31 @@ if (!isset($_SESSION['name'])) {
 
 $origin = $_SESSION['name'];
 
+// My PHP session only stores the username, which is sometimes less useful to me
+// than the actual name it is associated with... this grabs that name from the
+// database
 $query = "SELECT name FROM users WHERE username = '" . $origin . "'";
 $check = mysqli_query($conn2, $query);
+
+if ($check == FALSE) { 
+	die("could not execute statement $query<br />");
+}
+
 $originName = $check->fetch_array(MYSQLI_NUM);
 $name = $originName[0];
 
+// Lets say the laptop dies while a student is choose a destination - they'll no
+// longer be able to check in or out ... trapped, alone ... this fixes that,
+// cleaning things up everytime its associated page is ran
 $query = "DELETE FROM hallmonitor WHERE destination = '" . "TBD" . "' AND origin = '" . $name . "'";
 $check = mysqli_query($conn, $query);
 
+if ($check == FALSE) { 
+	die("could not execute statement $query<br />");
+}
+
+// There's probably an ID floating around in the session too, so lets make sure
+// that gets fixed
 if (isset($_SESSION["id"])) {
   unset($_SESSION["id"]);
 }
